backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
motor==3.3.2
pymongo==4.6.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-decouple==3.8
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.2.0
authlib==1.2.1
pydantic[email]>=2.8.0
itsdangerous==2.2.0
httpx==0.27.0
certifi==2023.11.17
openai==1.51.0
boto3>=1.34.0
botocore>=1.34.0
Pillow>=10.0.0
